pyproject.toml
[project]
name = "dutyflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openpyxl>=3.1.5",
    "pandas>=3.0.2",
    "pyyaml>=6.0.3",
]
